Refactor linter functionality into Validator class and update CLI to use the new structure
